[fix] 글쓰기 페이지 입력값들이 게시판 컴포넌트/게시판 글에 나타나도록 구현
