+++ src/App.js
@@ -16,9 +16,9 @@ import MemberDetail from './pages/MemberDetail';
 import MyPage from './pages/MyPage';
 import Edit from './pages/Edit';
 import Dues from './pages/Dues';
-import BoardNotice from './pages/BoardNotice';
-import BoardPosting from './pages/BoardPosting';
 import Board from './pages/Board';
+import BoardPosting from './pages/BoardPosting';
+import StudyBoard from './pages/StudyBoard';
 
 import { ThemeProvider } from 'styled-components';
 import theme from './styles/theme';
@@ -50,8 +50,8 @@ function App() {
           <Route path="/edit" element={<Edit />} />
           <Route path="/dues" element={<Dues />} />
           <Route path="/receipt" element={<Receipt />} />
+          <Route path="/StudyBoard" element={<StudyBoard />} />
           <Route path="/board" element={<Board />} />
-          <Route path="/notice" element={<BoardNotice />} />
           <Route path="/boardPosting" element={<BoardPosting />} />
         </Routes>
       </UserProvider>
