+++ src/components/home/HomeMain.jsx
@@ -145,7 +145,7 @@ const HomeMain = () => {
         </PenaltyItem>
         <BoardItem
           onClick={() => {
-            navi(`/notice`);
+            navi(`/board`);
           }}
         >
           게시판
