+++ src/pages/BoardPosting.jsx
@@ -17,7 +17,6 @@ const StyledText = styled.div`
   line-height: 19.09px;
 `;
 
-/* 임시 */
 const StyledTitle = styled.input`
   width: 88%;
   margin-top: 20px;
@@ -56,58 +55,54 @@ const StyledFileAttach = styled.div`
   margin-bottom: 148px;
 `;
 
-const BoardPosting = ({ initialBoardName, initialBoardContent }) => {
+const BoardPosting = ({ initialStudyName, initialStudyContent }) => {
   const navi = useNavigate();
-  const [boardName, setBoardName] = useState(initialBoardName);
-  const [boardContent, setBoardContent] = useState(initialBoardContent);
+  const [studyName, setStudyName] = useState(initialStudyName);
+  const [studyContent, setStudyContent] = useState(initialStudyContent);
   const [isCompleteEnabled, setIsCompleteEnabled] = useState(false);
 
-  const handleBoardNameChange = (e) => {
-    setBoardName(e.target.value);
+  const handleStudyNameChange = (e) => {
+    setStudyName(e.target.value);
   };
 
   const handleBoardContentChange = (e) => {
-    setBoardContent(e.target.value);
+    setStudyContent(e.target.value);
   };
 
-  /* const handleFocus = () => {
-    setIsKeyboardVisible(true);
-  }; 
-
-  /* const handleBlur = () => {
-    setIsKeyboardVisible(false);
-  }; */
-
-  // Board.jsx로 넘어가서 썼던 내용 보낼 수 있게
-  const handleCompleteClick = () => {
+  const handleBoardClick = () => {
     if (isCompleteEnabled) {
-      navi('/board', { state: { boardName, boardContent } }); // ?
+      navi('/board', {
+        state: {
+          studyName,
+          studyContent,
+        },
+      });
     }
   };
 
   useEffect(() => {
-    setIsCompleteEnabled(boardName && boardContent.length >= 1); // ?
-  }, [boardName, boardContent]);
+    setIsCompleteEnabled(studyName && studyContent.length >= 1);
+  }, [studyName, studyContent]);
 
   return (
     <StyledPosting>
       <PostingHeader
         isRightButtonEnabled={isCompleteEnabled}
-        onCompleteClick={handleCompleteClick}
+        onCompleteClick={handleBoardClick}
       />
       <StyledText>
         <StyledTitle
           type="text"
           placeholder="제목"
-          value={boardName}
-          onChange={handleBoardNameChange}
+          value={studyName}
+          onChange={handleStudyNameChange}
         />
       </StyledText>
       <StyledLine />
       <StyledText>
         <StyledContent
           placeholder="내용을 입력하세요."
-          value={boardContent}
+          value={studyContent}
           onChange={handleBoardContentChange}
         />
       </StyledText>
@@ -119,13 +114,13 @@ const BoardPosting = ({ initialBoardName, initialBoardContent }) => {
 };
 
 BoardPosting.propTypes = {
-  initialBoardName: PropTypes.string,
-  initialBoardContent: PropTypes.string,
+  initialStudyName: PropTypes.string,
+  initialStudyContent: PropTypes.string,
 };
 
 BoardPosting.defaultProps = {
-  initialBoardName: '',
-  initialBoardContent: '',
+  initialStudyName: '',
+  initialStudyContent: '',
 };
 
 export default BoardPosting;
